+++ Opcodes/ugsc.c
@@ -310,8 +310,8 @@ static int phaser1set(CSOUND *csound, PHASER1 *p)
     int   loop = (int) MYFLT2LONG(*p->iorder);
     int32  nBytes = (int32) loop * (int32) sizeof(MYFLT);
 
-    if (*p->istor == FL(0.0) || p->auxx.auxp == NULL || p->auxx.size<nBytes ||
-                                p->auxy.auxp == NULL || p->auxy.size<nBytes) {
+    if (*p->istor == FL(0.0) || p->auxx.auxp == NULL || (int)p->auxx.size<nBytes ||
+                                p->auxy.auxp == NULL || (int)p->auxy.size<nBytes) {
       csound->AuxAlloc(csound, nBytes, &p->auxx);
       csound->AuxAlloc(csound, nBytes, &p->auxy);
       p->xnm1 = (MYFLT *) p->auxx.auxp;
